update controller and queries functions

+++ API/Controllers/CommentController.cs
@@ -26,14 +26,18 @@ namespace API.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var comments = await _commentRepo.GetAllCommentAsync();
             var commentDto = comments.Select(x => x.ToCommentDto());
             return Ok(commentDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var comment = await _commentRepo.GetCommentByIdAsync(id);
             if (comment == null)
             {
@@ -45,18 +49,37 @@ namespace API.Controllers
         [HttpPost("{stockId}")]
         public async Task<IActionResult> Create([FromBody] CreateCommentDto commentDto, int stockId)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             if (!await _stockRepo.StockExists(stockId))
             {
                 return BadRequest("Stock doens't exist");
             }
-            var commentModel = commentDto.ToCommentFromCreateDto(stockId);
-            await _commentRepo.CreateCommentAsync(commentModel);
+            var commentModel = await _commentRepo.CreateCommentAsync(commentDto.ToCommentFromCreateDto(stockId));
             return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.ToCommentDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpPut("{id}")]
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto updateDto)
+        {
+            if (!ModelState.IsValid) return BadRequest();
+
+            var comment = await _commentRepo.UpdateCommentAsync(id, updateDto.ToCommentFromUpdateDto());
+            if (comment == null)
+            {
+                return NotFound();
+            }
+            return Ok(comment.ToCommentDto());
+
+        }
+
+
+
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var comment = await _commentRepo.DeleteCommentAsync(id);
             if (comment == null)
             {
