+++ API/Controllers/StockController.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using API.Data;
 using API.Dtos.Stock;
+using API.Helpers;
 using API.Interfaces;
 using API.Mappers;
 using Microsoft.AspNetCore.Mvc;
@@ -26,17 +27,21 @@ namespace API.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        public async Task<IActionResult> GetAll([FromQuery] ObjectQuery query)
         {
-            var stocks = await _stockRepo.GetAllStocksAsync();
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
+            var stocks = await _stockRepo.GetAllStocksAsync(query);
             var stockDto = stocks.Select(x => x.ToStockDto());
             return Ok(stockDto);
 
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var stock = await _stockRepo.GetStockByIdAsync(id);
             if (stock == null)
             {
@@ -48,17 +53,21 @@ namespace API.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var stockModel = stockDto.ToStockFromCreateDto();
             await _stockRepo.CreateStockAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         //You can pass only one attribute if prefer: [HttpPut("{id}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockDto updateDto)
         {
-            var stock = await _stockRepo.UpdateStockAsync(id, updateDto);
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
+            var stock = await _stockRepo.UpdateStockAsync(id, updateDto.ToStockFromUpdateDto());
             if (stock == null)
             {
                 return NotFound();
@@ -66,9 +75,11 @@ namespace API.Controllers
             return Ok(stock.ToStockDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var stock = await _stockRepo.DeleteStockAsync(id);
             if (stock == null)
             {
